+++ libs/wire-api/src/Wire/API/Team/Feature.hs
@@ -30,6 +30,7 @@ module Wire.API.Team.Feature
     TeamFeatureStatusNoConfig (..),
     TeamFeatureStatusWithConfig (..),
     HasDeprecatedFeatureName (..),
+    AllFeatureConfigs (..),
     defaultAppLockStatus,
     defaultClassifiedDomains,
 
@@ -44,6 +45,8 @@ module Wire.API.Team.Feature
   )
 where
 
+import Control.Lens.Combinators (dimap)
+import qualified Data.Aeson as Aeson
 import qualified Data.Attoparsec.ByteString as Parser
 import Data.ByteString.Conversion (FromByteString (..), ToByteString (..), toByteString')
 import Data.Domain (Domain)
@@ -63,15 +66,27 @@ import Wire.API.Arbitrary (Arbitrary, GenericUniform (..))
 ----------------------------------------------------------------------
 -- TeamFeatureName
 
--- | If you add a constructor here, you need to visit (at least) 4 places that are not caught
--- by ghc errors or test failures:
+-- | If you add a constructor here, you need extend multiple defintions, which
+--   aren't checked by GHC.
 --
--- * libs/wire-api/test/unit/Test/Wire/API/Roundtrip/Aeson.hs (calls to 'testRoundTrip')
--- * services/galley/src/Galley/API/Internal.hs (add a field to the 'InternalApi routes' record)
--- * libs/wire-api/src/Wire/API/Routes/Public/Galley.hs (add a field to the 'Api routes' record)
--- * services/galley/src/Galley/API/Teams/Features.hs (calls to 'getStatus')
--- * services/galley/schema/src/ (add a migration like the one in "V43_TeamFeatureDigitalSignatures.hs")
--- * services/galley/test/integration/API/Teams/Feature.hs (add an integration test)
+--   Follow this Checklist:
+--
+-- * libs/wire-api/test/unit/Test/Wire/API/Roundtrip/Aeson.hs
+--   * add call to 'testRoundTrip'
+-- * services/galley/src/Galley/API/Internal.hs
+--   * add a field to the 'InternalApi routes' record)
+-- * libs/wire-api/src/Wire/API/Routes/Public/Galley.hs
+--   * add a GET (and possible PUT) route with name prefix teamFeature<FEATURE_NAME>
+--   * add a GET route with name prefix featureConfig<FEATURE_NAME>
+-- * services/galley/src/Galley/API/Teams/Features.hs
+--   * extend getAllFeatureConfigs
+--   * extend getAllFeatures
+-- * services/galley/schema/src/
+--   * add a migration like the one in "V43_TeamFeatureDigitalSignatures.hs"
+-- * services/galley/test/integration/API/Teams/Feature.hs
+--   * add an integration test for the feature
+--   * extend testAllFeatures
+--
 --
 -- An overview of places to change (including compiler errors and failing tests) can be found
 -- in eg. https://github.com/wireapp/wire-server/pull/1652.
@@ -156,6 +171,8 @@ instance FromByteString TeamFeatureName where
         Right "classifiedDomains" -> pure TeamFeatureClassifiedDomains
         Right t -> fail $ "Invalid TeamFeatureName: " <> T.unpack t
 
+-- TODO: how do we make this consistent with 'KnownTeamFeatureNameSymbol'?  add a test for
+-- that?  anyway do we really need both?
 instance ToByteString TeamFeatureName where
   builder TeamFeatureLegalHold = "legalhold"
   builder TeamFeatureSSO = "sso"
@@ -375,3 +392,12 @@ data LowerCaseFirst
 instance StringModifier LowerCaseFirst where
   getStringModifier (x : xs) = toLower x : xs
   getStringModifier [] = []
+
+newtype AllFeatureConfigs = AllFeatureConfigs {_allFeatureConfigs :: Aeson.Object}
+  deriving stock (Eq, Show)
+  deriving (FromJSON, ToJSON, S.ToSchema) via (Schema AllFeatureConfigs)
+
+instance ToSchema AllFeatureConfigs where
+  schema =
+    named "AllFeatureConfigs" $
+      dimap _allFeatureConfigs AllFeatureConfigs jsonObject